Add all user configurable config opts to diagnose
With the default options in place, now we can iterate through the
configuration object data to print a more complete diagnose report.

The first option was to iterate through the constants defined in config
map, but that would give us empty values in the diagnose report.

I decided to iterate through the default values and ignore those that
the user cannot set as, that's what we do in the other integrations.