Add sources to the report output
The diagnose report output now contains information about which
source provides a given configuration value (when that source
is not `default`), and which other sources, if any, also provide
the given configuration value.

This brings the diagnose report output's behaviour regarding
sources in line with that of the Ruby and Elixir integration.