Remove `log_file_path` from diagnose config
This commit removes the `log_file_path` entry from the diagnose
config report and the corresponding output. The corresponding
private environment variable `_APPSIGNAL_LOG_FILE_PATH` is still
created, and the config has a `logFilePath` getter that can be
accessed by other parts of the code, but it is no longer part of
the "actual config", the one that is sent on the diagnose report.