Standardize key validation request
Add the other query params to the auth request.

Update the error message to be standard for all integrations. Not all
integrations necessarily have the response status code on error, so
make sure all integrations start with "Failed to validate".

I removed the empty JSON body from the request, because we don't parse
the body in this request and our test server interpreted it as the
parameter `{}` with the value of an empty String.

Add test for diagnose Push API key output and report.
Update the diagnose tests with tests for the Push API key validation in
the diagnose CLI.