Add tests for config module
Reorganize the tests to test single behavior. The "Private environment
variables" describe block also tested things for the `logFilePath`
option. Now it only tests if the private env vars are written if we
specify all the options.

I removed the usage of `process.env` to configure AppSignal, and rely on
the config options given to the Configuration module. We will test the
if the config is read from the environment variables in other tests.

I removed the default config initialization at the top-level
`beforeEach`, because it sets env vars that aren't overwritten if unset.
I could also call `resetEnv`, but I think in this case it's better to
only initialize it when needed in the configuration that is needed.
Otherwise we end up three levels deep calling `new Configuration(...)`
and `resetEnv()`.