Add requestHeaders allow list feature
`requestHeaders` allow list is now implemented. Now users are able to
define a list of headers they want to be shown in AppSignal. All
integrations except from Express make use of the core HTTP
instrumentation.

As we don't have helpers yet to access the Client stored in global
object, the code is repeated in HTTP instrumentation, and in Express
integration. This will be removed in favor of global helpers in the
future.