Rename apiKey config option to pushApiKey
Rename the `apiKey` config option to `pushApiKey` to match the naming in
the other integrations (Ruby & Elixir).

If the `apiKey` config option is set (which will be the case for
everyone upgrading) the `pushApiKey` config option will be set with the
`apiKey` value.

It does only for the `apiKey` option given to the constructor, the env
var is mapped to the new key by the config keys map.