Set log output to stdout on Heroku hosts
When the host is detected as a Heroku dyno (env var DYNO is set), we
default the log output to "stdout". We have the same behavior in our
other integrations.

Heroku has its own log drain to read all the app logs, and it's
difficult to access the `appsignal.log` file on dyno's, as we've noticed
for the diagnose report as well.

I've not made an exception for Dokku, as we've recently done for the
host metrics. Dokku also sets a DYNO env var and thus the log will
default to "stdout" on Dokku too. We've had no reports about problems
with this in the past in our other integrations.

Note that this config option only applies to the process output and
extension logger. The agent will always log to a file as it's a separate
process.

Fixes #495