Fix diagnose log_dir_path value
Use the `logFilePath`'s value as a base and get the parent directory
path to check.

This way the `log_dir_path` and `appsignal.log` paths don't point at
different directories. This could have happen because the `logFilePath`
getter (used by `appsignal.log`) does a permission check and falls back
on the system `/tmp` dir if the user configured path can't be written
to.