Remove default logPath option value
Remove the default value of the logPath option and only fall back on
the system tmp directory if no value is set. This way it matches its
behavior with the other integrations.

This will also make it easier to distinguish between user configured
paths and fallback paths in the future when we check if the fallback
directory is writable or not.

Fixes #510