Check if logPath fallback is writable
When the configured logPath is not writable, the integration falls back
on the system tmp dir. This commit checks if that directory is also
writable by the app process or not. We've seen this happen before in the
Ruby integration, so let's add the same check here.

This change means that the `logFilePath` getter can sometimes return an
`undefined` value. The integration will then not set the
`_APPSIGNAL_LOG_FILE_PATH` env variable, as it would point to a path
that can't be written to. In this case the agent will try to find a log
directory on its own, which will probably not succeed either and it will
shut down.