Revert: Update package lock on mono publish (#523)
This reverts commit 13c339ed7b3da5c8b795f29e9a510f5af714affd.

It broke the publish process, as it tried to fetch the unpublished
versions from the npm registry. It also wouldn't work in retrospect,
because the changes to the repository are committed already one step
earlier, so the publish process would be left with a dirty local state.

I have no other fix than manually committing it now.

[skip ci]