Unset the default logLevel in the integration
If the logLevel was set by default by the Node.js integration it would
always use that level, even if `debug` was set. This option is
deprecated, but if set, and `logLevel` is not configured by the app, we
should listen to these options.

I've removed the default `logLevel` from the integration, but it will be
set by the extension and agent's log_level behavior. Since there's no
centralized logger in the Node.js integration yet, we do not need
to update any configuration for the integration's logger.