Fix enableMinutelyProbes behaviour
Before this change, setting `enableMinutelyProbes` to `false`
would not disable the minutely probes system, it merely does not
register the default minutely probe.

The flag's behaviour has been changed so that the minutely probes
system is disabled in its entirety when it is set to `false`. The
`Probes` object within `Metrics` is replaced by a `NoopProbes`
object, which does nothing when a probe is registered.