Merge extension package into core one
The extension package is no longer an independent entity, it is now
merged into @appsignal/nodejs.

The installation script for the extension no longer aborts `npm install`
if something goes wrong during the process. This allows the users to
keep running their apps.

There's no need to use the script that creates the version file as
the extension constants are always available now.
So the script has been removed.