Fix probes not being overwritten
In the documentation for other integrations, we describe probes as
being overwritten if a different probe with the same name is
registered. While this was not documented as such for the Node.js
integration in our user-facing documentation, our internal
documentation for the `register` method describes its behaviour as
such.

However, since the `EventEmitter` listeners were not being removed,
when a new probe function was registered under the same name as an
existing probe function, both probe functions would be called
every minute. This commit fixes this, so that the behaviour matches
the other integrations and the internal documentation.