Avoid setting active to true in diagnose
Do not hardcode `active` to `true` when initialising a
configuration object for the diagnose. The value passed to the
extension object is still hardcoded to `true`, as this is necessary
for the extension to start, which is in turn needed to be able
to obtain diagnostic values from the extension.