Update report path to match new package structure
In PR #545 we've merged the nodejs-ext package into the nodejs package.
This change moved some files around, and moved the extension install
scripts one level deeper into the `extension` subdirectory. Update the
report path name generation to match the new depth of the file.

Previously: `packages/nodejs-ext/scripts/extension.js`
New: `packages/nodejs/scripts/extension/extension.js`

It's interesting that no tests failed because of this. I think it's not
caught in the diagnose_tests, because we stub the install report.

I've now added a simple test to test the presence of the report, and not
a parsing error that occurs when no file could be read or there was a
problem reading the file. I'm struggling with Jest's mock system to
write a test for this and I'd rather look at it when addressing issue
#550 that updates the path location, which makes it easier to stub I
hope.