Update install report location to ext directory
In PR #545 we've merged the nodejs-ext package into the nodejs package.
It no longer fails the package install on extension installation
failure, which means the ext directory remains after a installation
failure.

We can now store the install report file in the package's `ext`
directory, like we do in all the other integrations. This means we no
longer need to generate a filename for the install report to store it in
the system tmp directory.

Closes #550