Fix the CI build to include intentional failures
In the package mergers of PR #545, we merged the `nodejs-ext` package in
the `nodejs` package.

The tests that belong to the `nodejs-ext` package that tested the
behavior when the extension failed to install were not run after merging
that PR.

I've updated the build matrix to run those tests for the `nodejs`
package now.

For Jest I also had to configure a new root directory. In the
`nodejs-ext` package this was not configure, so I assume it previously
searched in all subdirectories.

For the clean command I also had to add the `build` directory be cleaned
up, otherwise it would continue to load the loaded extension without
problem.

[skip changeset]