Remove cpu and os fields from package.json
When the `cpu` or `os` fields are configured npm will not install the
package if the host CPU architecture or Operating System does not match
any of the configured values.

This previously meant the `nodejs-ext` package would not install and be
removed from the file system, which it also did if the extension
installation failed later on in the process. (It doesn't do that
anymore.) Since we merged the `nodejs-ext` package in the `nodejs`
package, these fields will cause serious problems with the integration.

If the nodejs package would fail to install the whole integration would
break. The app and other packages would try to call parts of the package
that is not present.

Remove the `os` and `cpu` fields to not restrict installation to any of
the configured CPU architectures or Operating Systems. Our extension
installer script figures out for which it can install the extension, and
write a report on failure. The `nodejs` package should always install
without issue.