Update package.json to link to mono repo paths (#563)
Following the `package.json` definition, we can specify where in the
repository the package is located. This will make commands like `npm
repo @appsignal/nodejs` link to the sub directory directly. And
hopefully the repository will be listed correctly on npmjs.org with this
change.

We can also look at configuring things like the homepage, contributors
list and other useful metadata if this works.

Because I've updated all `package.json` files, all packages have a
changeset for it.

https://docs.npmjs.com/cli/v8/configuring-npm/package-json#repository