Check if the correct package version is installed
In the `install_test_example_packages` script, check the output
of `npm list` if the installed packages are actually installed.
I've added this check, because running `npm install` won't fail if it
can't install a specific version if it already has installed something
(listing the package in the `package.json`).

It skips the `latest` version, because it won't know what version will
actually be installed.

Note that the `install_packages` script is not actually used at the
moment, because we don't run the unit tests against specific versions of
packages. Only the integration tests. There are no unit tests for
specific integration packages. If we decide to run future unit tests
against different versions, at least we won't run into the same issue
again.

Fixes #584
[skip changeset]