Update Span type used in ScopeManager (#587)
We changed the types used in PR #451. We no longer rely on the
`@appsignal/types` package for this, but define it in the package
itself.

The ScopeManager appears to not have been updated to match and is the
only place in which the `NodeSpan` from the types package is still
used. If we update this usage we can remove all Node.js types from the
types package.