Test the Transmitter class
Adds tests for the existing functionality of the `Transmitter`
class. It checks that the query parameters are added to the URL,
using values from the config, that the promise is resolved with a
status code and body, or rejected with an error, that the body is
parsed as JSON or replaced with an empty object, and that it uses
the CA file defined in the config in HTTPS requests.

Fixes an issue with the Transmitter where the full URL was used in
place of the path -- that is, `GET http://foo.com/bar` was sent
over the wire, instead of `GET /bar`.

Removes the unused `active` parameter from the Transmitter class,
and no longer passes it as `true` to the Config. Using the
Transmitter should not initialise the extension.

Changes the Transmitter implementation so that, when a CA file
could not be read, the `ca` property of the HTTPS options object
is unset, instead of being set to an empty string.