Test the extension download function
Modify the extension download script to allow for the functions
defined within it to be tested independently.

Test the extension download function `.downloadWithMirror`,
checking that it writes the downloaded contents to a stream, or
rejects the promise with an error if it receives a failure status
code.

Fix a bug in the extension download, where the URL printed when
reporting a failure to download from a given mirror would be
printed as `https:/...`, with only one slash, due to the misuse of
`path.join` to join components of the URL.

Amend the extension download logic so that it does not create a
file before knowing if the status code of the download request was
successful, and to avoid potentially piping data to the output
file stream before it is ready.

Also amend the extension to reject the promise when failing to
create the file, or when the connection fails to take place, and
test those scenarios.