Use transmitter to download the agent
Refactor the `Transmitter` class and create a new `.download`
method in the transmitter, to be used in the download request in
the extension installation script.

Create a new `.request` method, which is used by both `.download`
and `transmit`, which takes care of setting up a well-formed
request and using the CA certificate specified in the environment,
leaving the handling of the response stream to the caller.

Move the existing logic in the extension installation script that
relates to the initialisation of the HTTPS request to the
Transmitter.