Add tests for http outgoing integration
Add a SpanTestRegistry to store all Spans that are created during the
test. This registry will be accessible after the test to fetch all
created spans. It also mocks out the `Span.close` function so we can
call `toObject` on them later, and the Extension doesn't return an empty
String for the Span.

The mock for RootSpan and ChildSpan are mostly non destructive. Only
their `close` method is replaced with an implementation that doesn't
actually close the Span. Nothing that gets stored on the Span upon close
can be fetched from the extension if it's closed. To track if it was
closed or not we can check the `SpanTestRegistry.closedSpans` array.

## Types

I'm unsure about whole `(this as unknown) as Span` dance. It's the only
way I got it work. Some insight in how that could be done better is
welcome.

## About this implementation

This implementation is much the same as the Ruby gem and does not rely
on mock objects (`MockTracer`, `MockSpan`) that may or may not mimic
their real world behavior.

To avoid all the hassle of this Span creation tracking, because this was
not easy to figure out in Node.js, I'm thinking adding a test mode to
the extension where we can just ask which spans it created during the
test. That way we also only have to implement this once for all
integrations.
https://github.com/appsignal/appsignal-agent/issues/744

[skip changeset]