Keep Span object state before closing
After a span is closed, the extension forgets all about it. This
means that asking the extension for the JSON description of the
span won't work, and this is the underlying functionality for the
Span's `.toObject` method.

To work around this, the "tracked spans" created by the mock of
the Span module would mock out the `.close` method of the Span.
Instead of doing that, this commit saves the state returned by
`.toObject` before `.close` is called and the span is closed, and
returns that state in future calls to the `.toObject` method.