Ignore script test files in published package
When we publish a new version of the `@appsignal/nodejs` package the
test files for the extension install script were also included.
Some time ago there were no tests for this so it wasn't an issue, but
now that we have tests we need to explicitly ignore them in `.npmignore`
to not package them in the new published versions.

The output of `npm pack` before this change:

```
822B    scripts/extension/__tests__/extension.failure.test.js
3.9kB   scripts/extension/__tests__/extension.test.js
3.8kB   scripts/extension/__tests__/report.test.js
7.1kB   scripts/extension/extension.js
2.0kB   scripts/extension/report.js
2.5kB   scripts/extension/support/constants.js
1.3kB   scripts/extension/support/helpers.js
```

The output of `npm pack` after this change:

```
7.1kB   scripts/extension/extension.js
2.0kB   scripts/extension/report.js
2.5kB   scripts/extension/support/constants.js
1.3kB   scripts/extension/support/helpers.js
```

Fixes #596