Show extension load error on require
Before this commit, any errors in loading the extension wrapper
module wouldn't be shown if the `Extension` object wasn't started,
and the `BaseClient` object would not attempt to start the
extension if the extension wrapper wasn't loaded, meaning that
errors in loading the extension wrapper were, in practice, never
shown.

This commit changes those errors to be emitted when the extension
wrapper module is required, instead of when the `Extension` object
is started.