Fix tests on client start
These tests would assert the opposite behaviour of that which they
claimed to test. The test itself was broken, as it was spying on
a previously created client object, then creating a new client
object and performing actions on it, [as pointed out by @tombruijn
in #612][comment].

This commit fixes the existing tests so they perform the right
assertions, and adds a new one for the behaviour implied by the
tests.

[comment]: https://github.com/appsignal/appsignal-nodejs/pull/612#discussion_r804777513