Add central logger using Winston
This commits adds a new central logger using Winston. It is set up when
the client initializes. It will fallback to stdout if the provided log
file path is not accessible.

There's a new static method in client that gives access to the logger
through the globally stored client after initialization.