Fix Jest linter warnings and errors
Re-enable the following linter warnings and errors, fixing all
issues detected:

- `jest/expect-expect`: In tests where a function is merely called
to ensure calling it does not throw an error, wrap the call in an
explicit `expect(() => call()).not.toThrow()`. To allow `expect`
blocks to be reused across tests, configure the linter so that any
function whose name begins with `expect` is considered an expect.

- `jest/no-done-callback`: In tests where the usage of the `done`
callback is superfluous, remove it. In tests where it is called at
the end of a callback, as an implicit assertion that the callback
is called, wrap the callback in a `jest.fn()` and `expect` it to
be called.

- `jest/no-identical-title`: Ensure different tests within the
same describe block have different names. The only occurrence of
this was a copy-paste mistake.