Use ESLint
Add the ESLint linter to the project, along with the Typescript,
Prettier and Jest plugins for it, and configure ESLint to use them.

Replace uses of Prettier in the `lint` and `lint:write` actions
and in the `lint-staged` pre-commit hook with ESLint.

Configure ESLint so that all rules that currently trigger errors
or warnings are turned off. Subsequent commits will re-enable these
rules along with the necessary fixes to the code.