Use comments to disable rules when used
For rules which are infringed upon sparingly, use a comment to
disable the rule at the point in the code where it is infringed,
instead of disabling the rule for the entire codebase.

This helps to ensure that we do not accidentally add any new code
that infringes the rule.

Re-enable the `@typescript-eslint/no-this-alias`,
`@typescript-eslint/ban-ts-comment` and
`@typescript-eslint/triple-slash-reference` linter rules.