Allow nested values in setSampleData
Fix the signature of setSampleData so that it accepts any array or
object, instead of only arrays or objects with up to two levels of
nesting. This allows the user to pass arbitrary objects to sample
data properties that accept them, such as `custom_data`.

Note that not all sample data properties accept nested values.