Always print the install report in CI (#628)
Makes debugging easier to see what's going on, even when it doesn't
think it went wrong but it did.

Move it from the epilogue to the last command in the job. If it's in the
epilogue and the file is missing it doesn't fail the task, which it
should. Otherwise we don't know that the install did not trigger.