Delete job specific cache in CI jobs (#635)
To move build artifacts around from job to job in the same workflow,
make sure it uses the artifacts generated by this workflow's run.

Caches are not overwritten if it already exists. Make sure to always
remove the cache if it exists. Otherwise a retry of a build may to
continue to use the cache from the first time the workflow was run,
which is not what we want and may introduce silent bugs.

I did not choose to use artifacts here as artifacts may incur an
addition charge later on. By deleting the cache we get the same
behavior.

Closes #631