Fix `datamap.setFloat` implementation
The existing implementation for `datamap.setFloat` mistakenly
attempted to use the second argument as both the value to add to
the map and the map to add it to. This resulted in an "Invalid
argument" error thrown when trying to cast the value as a map.

The tests for `Data.generate` tested sending floating point values
to the agent using the value `99.0`. Since JavaScript has a unified
`number` type for integers and floats alike, `99.0` is the same
number as `99`.

The implementation of `Data.generate` decides whether to call
`datamap.setInteger` or `datamap.setFloat` based on whether
`Number.isInteger` returns `true` or `false` -- and it will return
`true` for `99.0`, as it would for `99`. This means that, in
practice, `datamap.setFloat` was not covered by the tests.