Improve ScopeManager tests
Test ScopeManager's enable/disable methods. These methods were
not being tested, other than their implementation not throwing an
error when called. The tests now check that the underlying async
hook is enabled/disabled.

Fix the `withContext` test for restoring the scopes. Before, it
was checking that the active span and the root span were equal --
which was true, but only because they were both `undefined`. The
equivalent test now makes assertions about the behaviour of the
method as a whole, checking that both active and root spans are
restored.

Always use `toBe` to compare spans. Spans are identified by a
reference empty object which the AppSignal agent keeps track of.
These empty objects are always strict equal to each other, so
Span objects should be compared to each other with identity
equality (`Object.is`).