Fix Apollo server error reporting
Ensure `setRoot` sets an active scope. If there is no active
scope, calling `scopeManager.setRoot` will set the given root
span as both the active span and the root span for the current
process. This fixes a bug where processes lose track of their
root span after `tracer.withSpan` is called.

Remove a check in the HTTP outgoing instrumentation that checks
if there is a current span before creating a root span.

Add integration tests for the current behaviour of the Apollo
Server integration. It can report a successful query, including
the query body, and report errors that take place during the
parsing and validation stages.

The error reporting was previously not working due to the bug
described above regarding `scopeManager.setRoot`.