Use `host` if `hostname` is undefined
When processing request options in our outgoing HTTP request
instrumentation, if `hostname` is undefined, use `host` as the
host name instead.

Note that the meanings for `hostname` and `host` differ between
the request options object in the Node.js `http` library and the
`url.URL` object from the standard WHATWG URL API.

In `url.URL` objects, `hostname` is the host name ("google.com")
and `host` is the host name and port ("google.com:80").

In the `http` library, `host` is expected to be the host name,
without the port. However, for compatibility with `url.parse()`
from the `url.URL` API, `hostname` is supported as an alias for
`host`, overriding it if both are present.