Check if extension installed successfully on CI
We printed the contents of the install report, but did not check if it
the result status was successful or not. This then lead to all the
specific tests to be run in an integration without an extension.

This will halt the CI on the build step rather than the tests for all
the supported libraries. Saves time and energy for us and the CI, and
gives a clearer picture in the CI overview what went wrong.