Move condition to build the package to a script
For local development we want to build the package before running
install, this will save hassle with fresh clones of the repository and
on `mono clean`/`npm run clean`.

The check we did was a Unix only check within the if-statement with
`test`. Move the check and execution of `npm run build` to a new Node.js
file so it runs on any Operating System.

Fixes #646