Add Node 18 to build matrix
Node 18 is now available and Node 12 is in EOL. This commits adds Node
18 to the build matrix and removes Node 12.

As Node 18 requires glibc 2.28 or newer, the matrix needs to run on at
least Ubuntu 20.04, so it's beend updated.

We don't need to specify C 8 anymore for any build as all versions work
by default with the version included in the new OS image.