Fix span tests after field rename in span_to_json
The fields returned for the start time have been renamed and the
nanoseconds field has been added. Update the tests to match.

I haven't implemented the nanoseconds one because it returns negative
numbers, something which should not be possible. It appears to be a
problem in our Node.js library time conversion.