Add createRootSpan function (#656)
By using the createRootSpan function in the HTTP incoming
instrumentation, we ensure ensure that every request always create a new
RootSpan. Between the requests in express, the context could be the same
(having the same executionAsyncId) as the previous request.

The problem that could occur (especially in our new OpenTelemetry
instrumentation), is that spans from new requests could continue to be
tracked on the RootSpan of the previous request. Which could cause us to
missout on these requests. Or, in the case of the OpenTelemetry
instrumentation, create a new RootSpan for the child span, because the
`spanId` and `traceId` would be empty, because the RootSpan was already
closed.

We removed the if-statement in the scope manager. When creating a new
RootSpan for a context that already has a RootSpan, with the
if-statement intact, the `rootSpan` function would return the new
RootSpan, but `currentSpan` function would return the old RootSpan. We
need to update the scopes in the ScopeManager to also set the new
RootSpan on the "scopes" for the context, even if it's already set
before.