Fix instrument options always returning true
The `|| true` fallback for the default value would always set the
options to true for the `initCorePlugins` function.

Allow the value to be `undefined`, which is necessary because of the
`Partial<AppsignalOptions>` argument to the client constructor.

We only check if the `instrumentHttp` option value is `false` and not if
it's true (the default). If a user explicitly sets it to `false` the
instrumentation will not be enabled.