Add options for disabling default instrumentation (#664)
* Add options for disabling default instrumentation

We previously had the undocumented configuration option
`ignoreInstrumentation` that didn't follow our previously set out naming
and behavior conventions for configuration options.

Split up this configuration option in separate configuration options for
the different instrumentations.

Closes #519

* Merge the instrumentHttp and https options

These two options basically do the same thing, instrument the http and
https modules, but there will most likely not be a scenario where one
would instrument http requests but not https requests or the other way
around. Merge both options into the `instrumentHttp` option so one
option disables both instrumentations.

* Fix config reading issue for instrument options

The client loaded the instrument options too early, before the config
was all read properly. This mean the `APPSIGNAL_INSTRUMENT_*` env vars
were ignored.

* Fix instrument options always returning true

The `|| true` fallback for the default value would always set the
options to true for the `initCorePlugins` function.

Allow the value to be `undefined`, which is necessary because of the
`Partial<AppsignalOptions>` argument to the client constructor.

We only check if the `instrumentHttp` option value is `false` and not if
it's true (the default). If a user explicitly sets it to `false` the
instrumentation will not be enabled.