Remove AppSignal requirement for SpanProcessor
For integrations like Express we'll rely on the OpenTelemetry
instrumentations to create root spans.

AppSignal instrumentation and OpenTelemetry instrumentation is not
compatible with each other this way. This change changes the requirement
from an AppSignal instrumentation to the OpenTelemetry instrumentation.

If not, the parentSpanId and traceId of the AppSignal span is empty and
it will create incidents for all the separate child spans.

Part of https://github.com/appsignal/opentelemetry/issues/14
Requires https://github.com/appsignal/appsignal-agent/pull/762