Lock apollo-server package versions in CI
Make sure that the apollo-server package uses the same version as the
apollo-server-plugin-base package.

The apollo-server package is configured as matching any version (`*`) in
the test app's `package.json`. This should install the right version.

The apollo-server and apollo-server-plugin-base packages do not follow
the same release scheme so that's why the versions do not appear to
match.

I've also updated the versions against we test to be the last three
ones, except the 3.8 release (3.7 was skipped).